Refactor `StormLaws` (#727)
In this PR I refactored `StormLaws` class with a way to test same `Producer` on both Storm and Memory platforms and compare their results. In future this approach can be extended to Scalding platform as well with a very unified way of testing.

Also I've included test for #725, where we build Storm topology which fails with `ClassCastException` in runtime.