Refactor `Storm` platform to introduce `Edge`s (#728)
In order to introduce grouped `leftJoin` I would like to make tuples we send between nodes to be more precise. In particular I want to send `(K, V)` tuples as tuples with `key` and `value` fields.

In this PR I did a small refactoring to introduce Edge trait which correspond to edges in Storms topology DAG - it contains a way how to serialize/deserialize data into `Fields` and how to group data over this edge.