Introduce `Topology` class (#731)
This PR introduces `Topology` class which corresponds to Storm's dag with nodes and edges.

We firstly build `Topology` instance out of planned `Dag` and then build `PlannedTopology` out of it. At the moment of topology building we verify that edges in topology are compatible with each other and output format on outcoming edges is the same.