Fix bug in ExpressionDag.fanOut
There was a subtle bug in ExpressionDag.fanOut which caused
non-idempotency, and underapplication of rules in some cases. The
problem was in computing fanOut in Id space, where there is actually no
1:1 function between Ids and N nodes. The solution is to compute fanOut
directly in N space which is what is meaningful anyway.

This seems to fix the bug even with rather large numbers of trials